even more c++ std lib compliance now.

+++ hcore/macro.hxx
@@ -87,7 +87,7 @@ Copyright:
 #endif /* not NDEBUG */
 /*! \brief Convinience macro to obtain number of elements of declared array.
  */
-#define countof( array ) ( sizeof ( yaal::YaalArrayElementCountHelper( ( array ) ) ) )
+#define countof( array ) ( static_cast<int long>( sizeof ( yaal::YaalArrayElementCountHelper( ( array ) ) ) ) )
 
 #endif /* not YAAL_HCORE_MACRO_HXX_INCLUDED */
